フォルダ構成修正、csv一覧取得の実装
