+++ Backend/internal/dto/auth-dto.go
@@ -1,0 +1,11 @@
+package dto
+
+type SignupInput struct {
+	Email    string `json:"email" binding:"required,email"`
+	Password string `json:"password" binding:"required,min=8"`
+}
+
+type LoginInput struct {
+	Email    string `json:"email" binding:"required,email"`
+	Password string `json:"password" binding:"required,min=8"`
+}
