+++ Backend/internal/models/users.go
@@ -1,0 +1,10 @@
+package models
+
+import "gorm.io/gorm"
+
+type User struct {
+	gorm.Model
+	Email      string `gorm:"uniqueIndex;not null"`
+	Password   string `gorm:"not null"`
+	ImportData []ImportData
+}
