+++ Backend/internal/repositories/user-repository.go
@@ -1,0 +1,37 @@
+package repositories
+
+import (
+	"main/internal/models"
+
+	"gorm.io/gorm"
+)
+
+type IUserRepository interface {
+	CreateUser(user *models.User) error
+	FindUser(email string) (*models.User, error)
+}
+
+type UserRepository struct {
+	db *gorm.DB
+}
+
+func NewUserRepository(db *gorm.DB) IUserRepository {
+	return &UserRepository{db: db}
+}
+
+func (r *UserRepository) CreateUser(user *models.User) error {
+	result := r.db.Create(&user)
+	if result.Error != nil {
+		return result.Error
+	}
+	return nil
+}
+
+func (r *UserRepository) FindUser(email string) (*models.User, error) {
+	var user models.User
+	result := r.db.Where("email = ?", email).First(&user)
+	if result.Error != nil {
+		return nil, result.Error
+	}
+	return &user, nil
+}
