--- Backend/utils/hash.go
@@ -1,21 +1,0 @@
-package utils
-
-import (
-	"golang.org/x/crypto/bcrypt"
-)
-
-func HashPassword(password string) (string, error) {
-	hashPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	if err != nil {
-		return "", err
-	}
-	return string(hashPassword), nil
-}
-
-func CheckPassword(hashPassword string, password string) error {
-	err := bcrypt.CompareHashAndPassword([]byte(hashPassword), []byte(password))
-	if err != nil {
-		return err
-	}
-	return nil
-}
